pyproject.toml
[project]
name = "noticebot"
version = "0.1.0"
description = "noticebot"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "飞书", module_name = "nonebot.adapters.feishu" }
]
plugins = []
plugin_dirs = []
builtin_plugins = []

[tool.poetry]
name = "noticebot"
version = "0.1.0"
description = ""
authors = ["jerry-harm <109275116+jerry-harm@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
nonebot-adapter-onebot = "^2.4.5"
nonebot2 = {extras = ["fastapi", "httpx"], version = "^2.3.3"}
lark-oapi = "^1.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
